requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
sqlalchemy-migrate
python-dotenv
